chore: extract manifest_dir function
